// Desktop/cs5500-gui-arch-huangweili97-main/src/types/types.tsx
export interface QuestionProps {
    qid: string;
    title: string;
    text: string;
    tagIds: string[];
    askedBy: string;
    askDate: Date;
    ansIds: string[];
    views: number;
    getAnswerCount: () => number;
    getQuestionViews: () => number;
    getTagsId: () => string[];
    calculateTimeElapsed: () => string;
}
// ...
export interface TagParamType {
    tid: string;
    name: string;
}